+++ libs/content/blog/project.json
@@ -4,7 +4,7 @@
   "projectType": "library",
   "targets": {
     "compile": {
-      "executor": "@chair-flight/core/blog:compile"
+      "executor": "@chair-flight/providers/blog:compile"
     }
   },
   "tags": []
