+++ libs/content/question-bank-atpl/project.json
@@ -4,14 +4,14 @@
   "projectType": "library",
   "targets": {
     "parseLosXlsx": {
-      "executor": "@chair-flight/core/question-bank:parseLosXlsx",
+      "executor": "@chair-flight/providers/question-bank:parseLosXlsx",
       "options": {
         "xlsxPath": "{projectRoot}/subjects/learning-objectives.xlsx",
         "outputJsonPath": "{projectRoot}/subjects/learning-objectives.json"
       }
     },
     "compile": {
-      "executor": "@chair-flight/core/question-bank:compile"
+      "executor": "@chair-flight/providers/question-bank:compile"
     }
   },
   "tags": []
