+++ libs/content/question-bank-type/project.json
@@ -4,7 +4,7 @@
   "projectType": "library",
   "targets": {
     "compile": {
-      "executor": "@chair-flight/core/question-bank:compile"
+      "executor": "@chair-flight/providers/question-bank:compile"
     }
   },
   "tags": []
