+++ libs/providers/analytics/.eslintrc.json
