+++ libs/providers/analytics/prisma/schema.prisma
