+++ libs/providers/analytics/tsconfig.json
