+++ libs/providers/blog/package.json
@@ -1,4 +1,4 @@
 {
-  "name": "core-blog",
+  "name": "providers-blog",
   "executors": "./executors.json"
 }
