+++ libs/providers/question-bank/package.json
@@ -1,4 +1,4 @@
 {
-  "name": "core-question-bank",
+  "name": "providers-question-bank",
   "executors": "./executors.json"
 }
