+++ libs/providers/question-bank/src/question-bank-docs-meta-schema.ts
