+++ libs/trpc/server/src/routers/common/analytics-router.ts
@@ -4,7 +4,7 @@ import {
   createTrackEvent,
   generateDailyCounts,
   getPageVisits,
-} from "@chair-flight/core/analytics";
+} from "@chair-flight/providers/analytics";
 import { publicProcedure, router } from "../../config/trpc";
 import type { PageEvent, SimplifiedTrackEvent } from "@chair-flight/base/types";
