+++ libs/trpc/server/src/routers/common/tests-router.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 import { makeMap } from "@chair-flight/base/utils";
 import { createTest, newTestConfigurationSchema } from "@chair-flight/core/app";
-import { questionBanks } from "@chair-flight/core/question-bank";
 import { questionBankNameSchema } from "@chair-flight/core/schemas";
+import { questionBanks } from "@chair-flight/providers/question-bank";
 import { publicProcedure, router } from "../../config/trpc";
 
 export const testsRouter = router({
