+++ libs/trpc/server/src/routers/containers/blog-router.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { blog } from "@chair-flight/core/blog";
+import { blog } from "@chair-flight/providers/blog";
 import { compileMdx } from "../../common/compile-mdx";
 import { publicProcedure, router } from "../../config/trpc";
