+++ libs/trpc/server/src/routers/containers/flashcards-router.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { getRandomShuffler } from "@chair-flight/base/utils";
-import { questionBanks } from "@chair-flight/core/question-bank";
 import { questionBankNameSchema } from "@chair-flight/core/schemas";
+import { questionBanks } from "@chair-flight/providers/question-bank";
 import { publicProcedure, router } from "../../config/trpc";
 
 export const flashcardsContainersRouter = router({
