+++ libs/trpc/server/src/routers/containers/overviews-router.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { questionBanks } from "@chair-flight/core/question-bank";
+import { questionBanks } from "@chair-flight/providers/question-bank";
 import { publicProcedure, router } from "../../config/trpc";
 
 export const overviewsContainersRouter = router({
