+++ libs/trpc/server/src/routers/pages/blog-pages-router.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { blog } from "@chair-flight/core/blog";
+import { blog } from "@chair-flight/providers/blog";
 import { publicProcedure, router } from "../../config/trpc";
 
 export const blogPagesRouter = router({
