+++ libs/trpc/server/src/routers/pages/modules-pages-router.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { questionBanks } from "@chair-flight/core/question-bank";
 import { questionBankResourceSchema } from "@chair-flight/core/schemas";
+import { questionBanks } from "@chair-flight/providers/question-bank";
 import { publicProcedure, router } from "../../config/trpc";
 import type { QuestionBankName } from "@chair-flight/base/types";
