+++ tsconfig.base.json
@@ -31,11 +31,11 @@
       "@chair-flight/base/errors": ["libs/base/errors/src/index.ts"],
       "@chair-flight/base/types": ["libs/base/types/src/index.ts"],
       "@chair-flight/base/utils": ["libs/base/utils/src/index.ts"],
-      "@chair-flight/core/analytics": ["libs/core/analytics/src/index.ts"],
+      "@chair-flight/providers/analytics": ["libs/providers/analytics/src/index.ts"],
+      "@chair-flight/providers/blog": ["libs/providers/blog/src/index.ts"],
+      "@chair-flight/providers/question-bank": ["libs/providers/question-bank/src/index.ts"],
       "@chair-flight/core/app": ["libs/core/app/src/index.ts"],
       "@chair-flight/core/github": ["libs/core/github/src/index.ts"],
-      "@chair-flight/core/question-bank": ["libs/core/question-bank/src/index.ts"],
-      "@chair-flight/core/blog": ["libs/core/blog/src/index.ts"],
       "@chair-flight/core/schemas": ["libs/core/schemas/src/index.ts"],
       "@chair-flight/react/analytics": ["libs/react/analytics/src/index.ts"],
       "@chair-flight/react/components": ["libs/react/components/src/index.ts"],
